+++ scraper/slotscraper/ordoclic/ordoclic.py
